Many signedness changes to work towards no -Wconversion warnings

+++ include/vb/Bitmap.h
@@ -76,7 +76,7 @@ namespace vb {
         Image img(H.title, {256, 256});
         img.show();
         for (int i = 0; i < 256; ++i)
-            for (int j = 0; j < 256; ++j) img.put({i, j}, Color(i, j, (8 * (i + j)) % 256));
+            for (int j = 0; j < 256; ++j) img.put({i, j}, Color(uint8_t(i), uint8_t(j), uint8_t((8 * (i + j)) % 256)));
         img.hide();
     }
 #endif
